+++ CHANGES.md
@@ -1,5 +1,12 @@
 # 変更点一覧
 
+## 2.9から2.10への変更点
+
+* RendererクラスでSmartyをロードしなくなりました。Appid_ControllerでSmartyをrequireするか、自前でオートロードしてください。
+* SOAP Gatewayを廃止しました。
+* Ethna_DB_PEARを廃止しました。
+* Ethna_Controllerでpreg_matchがPHP5.5でDEPRECATEDになるのを修正しました。
+
 ## 2.8から2.9への変更点
 
 * Backend#performを Controller#performへ引っ越ししました。